+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Changed
 - Load CSS and JS only when it's really needed
 - Osmium is working again, so it's back on the menu
-- Fitting detail page now modularized, that means you can now which information will be shown
+- Fitting detail page now modularized, that means you can now influence which information will be shown
 
 ## [v0.1](https://github.com/ppfeufer/eve-online-fitting-manager/releases/tag/v0.1) - 2017-08-02
 [Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.1-r20170730...v0.1)