+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
 
 ## [In Development](https://github.com/ppfeufer/eve-online-fitting-manager)
-[Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.1...HEAD)
+[Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.2...HEAD)
+- in development
+
+## [v0.2](https://github.com/ppfeufer/eve-online-fitting-manager/releases/tag/v0.2) - 2017-08-05
+[Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.1...v0.2)
 ### Added
 - Loading animation while waiting for market data